Код переписан ближе к паттерну mvc
