+++ backend/core.py
@@ -1,3 +1,27 @@
-import redis
+import logging
 
-caching = redis.Redis("192.168.1.9", "6379")
+import redis
+from fastapi import FastAPI
+from starlette.middleware.cors import CORSMiddleware
+from starlette.staticfiles import StaticFiles
+from starlette.templating import Jinja2Templates
+
+from config import origins, REDIS_HOST, REDIS_PORT
+
+
+app = FastAPI()
+logging.basicConfig(level=logging.INFO)
+logger = logging.getLogger(__name__)
+
+app.add_middleware(
+    CORSMiddleware,
+    allow_origins=origins,
+    allow_credentials=True,
+    allow_methods=["*"],
+    allow_headers=["*"],
+)
+
+
+templates = Jinja2Templates(directory="templates")
+app.mount("/static", StaticFiles(directory="static"), name="static")
+caching = redis.Redis(REDIS_HOST, REDIS_PORT)
